+++ test/unit/lib/Api.specs.js
@@ -11,6 +11,7 @@ test.describe('Api', () => {
   let usersMocks
   let servicesMocks
   let securityTokensMocks
+  let logMocks
   let api
 
   test.beforeEach(() => {
@@ -18,6 +19,7 @@ test.describe('Api', () => {
     usersMocks = new mocks.api.Users()
     securityTokensMocks = new mocks.api.SecurityTokens()
     servicesMocks = new mocks.api.Services()
+    logMocks = new mocks.api.Logs()
     api = Api(baseMocks.stubs.service)
   })
 
@@ -26,6 +28,7 @@ test.describe('Api', () => {
     usersMocks.restore()
     servicesMocks.restore()
     securityTokensMocks.restore()
+    logMocks.restore()
   })
 
   test.describe('instance', () => {
@@ -53,7 +56,9 @@ test.describe('Api', () => {
           'addAbility',
           'deleteAbility',
           'abilityAction',
-          'abilityState'
+          'abilityState',
+          'abilityEvent',
+          'getLogs'
         )
       })
     })