+++ test/unit/lib/models/Log.mocks.js
@@ -1,0 +1,32 @@
+
+const test = require('narval')
+
+const log = require('../../../../lib/models/log')
+
+const Mock = function () {
+  const sandbox = test.sinon.createSandbox()
+
+  const logStub = {
+    save: sandbox.stub().usingPromise().resolves()
+  }
+
+  const LogStub = sandbox.stub().returns(logStub)
+  LogStub.find = sandbox.stub().usingPromise().resolves()
+
+  const stubs = {
+    log: logStub,
+    Log: LogStub,
+    Model: sandbox.stub(log, 'Model').returns(LogStub)
+  }
+
+  const restore = function () {
+    sandbox.restore()
+  }
+
+  return {
+    stubs: stubs,
+    restore: restore
+  }
+}
+
+module.exports = Mock