+++ test/unit/mocks.js
@@ -14,6 +14,7 @@ const UsersApi = require('./lib/api/Users.mocks')
 const SecurityTokensApi = require('./lib/api/SecurityTokens.mocks')
 const ServicesApi = require('./lib/api/Services.mocks')
 const AbilitiesApi = require('./lib/api/Abilities.mocks')
+const LogsApi = require('./lib/api/Logs.mocks')
 
 const Commands = require('./lib/Commands.mocks')
 const UserCommands = require('./lib/commands/User.mocks')
@@ -21,12 +22,14 @@ const SecurityTokenCommands = require('./lib/commands/SecurityToken.mocks')
 const ComposedCommands = require('./lib/commands/Composed.mocks')
 const ServiceCommands = require('./lib/commands/Service.mocks')
 const AbilityCommands = require('./lib/commands/Ability.mocks')
+const LogCommands = require('./lib/commands/Log.mocks')
 
 const Models = require('./lib/Models.mocks')
 const UserModel = require('./lib/models/User.mocks')
 const ServiceModel = require('./lib/models/Service.mocks')
 const SecurityTokenModel = require('./lib/models/SecurityToken.mocks')
 const AbilityModel = require('./lib/models/Ability.mocks')
+const LogModel = require('./lib/models/Log.mocks')
 
 module.exports = {
   Api,
@@ -34,20 +37,23 @@ module.exports = {
     Users: UsersApi,
     SecurityTokens: SecurityTokensApi,
     Services: ServicesApi,
-    Abilities: AbilitiesApi
+    Abilities: AbilitiesApi,
+    Logs: LogsApi
   },
   commands: {
     User: UserCommands,
     SecurityToken: SecurityTokenCommands,
     Composed: ComposedCommands,
     Service: ServiceCommands,
-    Ability: AbilityCommands
+    Ability: AbilityCommands,
+    Log: LogCommands
   },
   models: {
     User: UserModel,
     SecurityToken: SecurityTokenModel,
     Service: ServiceModel,
-    Ability: AbilityModel
+    Ability: AbilityModel,
+    Log: LogModel
   },
   Base,
   Client,